test: address ISTFT oracle review comments
